optimizes the code
