+++ routes/api/contacts.js
@@ -1,84 +1,17 @@
 const express = require("express");
-const Joi = require("joi");
-
-const contacts = require("../../models/contacts");
-
-const { HttpError } = require("../../helpers");
 
 const router = express.Router();
 
-const addSchema = Joi.object({
-	name: Joi.string().required(),
-	email: Joi.string().email().required(),
-	phone: Joi.string()
-		.pattern(/^\(\d{3}\) \d{3}-\d{4}$/)
-		.required(),
-});
+const controllers = require("../../controllers/contacts");
 
-router.get("/", async (req, res, next) => {
-	try {
-		const result = await contacts.listContacts();
-		res.json(result);
-	} catch (error) {
-		next(error);
-	}
-});
+router.get("/", controllers.getAllContacts);
 
-router.get("/:contactId", async (req, res, next) => {
-	try {
-		const { contactId } = req.params;
-		const result = await contacts.getContactById(contactId);
-		if (!result) {
-			throw HttpError(404, "Not Found");
-		}
-		res.json(result);
-	} catch (error) {
-		next(error);
-	}
-});
+router.get("/:contactId", controllers.getById);
 
-router.post("/", async (req, res, next) => {
-	try {
-		const { error } = addSchema.validate(req.body);
-		console.log(error);
-		if (error) {
-			throw HttpError(400, "missing required name field");
-		}
-		const result = await contacts.addContact(req.body);
-		res.status(201).json(result);
-	} catch (error) {
-		next(error);
-	}
-});
+router.post("/", controllers.addContact);
 
-router.delete("/:contactId", async (req, res, next) => {
-	try {
-		const { contactId } = req.params;
-		const result = await contacts.removeContact(contactId, req.body);
-		if (!result) {
-			throw HttpError(404, "Not found");
-		}
-		res.json({ message: "contact deleted" });
-	} catch (error) {
-		next(error);
-	}
-});
+router.delete("/:contactId", controllers.deleteContact);
 
-router.put("/:contactId", async (req, res, next) => {
-	try {
-		const { error } = addSchema.validate(req.body);
-		if (error) {
-			throw HttpError(400, "missing fields");
-		}
-		const { contactId } = req.params;
-		const result = await contacts.updateContact(contactId, req.body);
-		if (!result) {
-			throw HttpError(404, "Not found");
-		}
-		res.json(result);
-	} catch (error) {
-		next(error);
-	}
-});
+router.put("/:contactId", controllers.changeContact);
 
 module.exports = router;
